requirements-arm64.txt
django
django-jinja
django-session-security
django-guardian
django-crispy-forms
django-huey
crispy-bootstrap5
psycopg2-binary
pydicom
pynetdicom
python-gdcm
numpy==1.26.4
scipy
pillow
pandas
openpyxl
tensorflow==2.15.0 # arm64 version
# torch         # installed via URL
# torchvision
gunicorn
pendulum
redis